WIP Started conversion to use Jobs containing Events
